+++ site/content/release-notes.md
@@ -27,10 +27,13 @@ title = "Release Notes"
   `org.eclipse.hono.service.AbstractProtocolAdapterBase.addProperties` methods have
   been changed to accept an additional parameter of type `TenantObject` which may contain default
   properties defined for the tenant to be included in downstream messages.
-* The `Get Registration Information` operation of the Device Registration API is not optional anymore, 
+* The *get Registration Information* operation of the Device Registration API is not optional anymore, 
   it is now mandatory to implement. For device registry implementations based on the
   `CompleteRegistrationService` interface, there is no change needed as the operation is already
   defined there.
+* The response message of the *assert Device Registration* operation does not contain an assertion
+  token anymore. The `org.eclipse.hono.service.registration.BaseRegistrationService` class
+  has been adapted accordingly.
 * The already deprecated `org.eclipse.hono.client.CommandConsumerFactory.closeCommandConsumer`
   method has been removed.